keystore2: Block key attestation for Google Play Services
In order to enforce SafetyNet security, Google Play Services is now
using hardware attestation for ctsProfile validation in all cases, even
when basic attestation is selected. The SafetyNet API response from GMS
will report that basic attestation was used, but under the hood,
hardware attestation is always used regardless of the reported state.
This results in SafetyNet failing to pass due to TrustZone reporting an
unlocked bootloader (and a partially invalidated root of trust) in the
key attestation result.

We can still take advantage of the fact that this usage of hardware
attestation is opportunistic - that is, it falls back to basic
attestation if key attestation fails to run - and prevent GMS from using
key attestation at the framework level. This causes it to gracefully
fall back to basic attestation and pass SafetyNet with an unlocked
bootloader.

Key attestation is still available for other apps, as there are valid
uses for it that do not involve SafetyNet.

The "not implemented" error code from keymaster is used to simulate the
most realistic failure condition to evade detection, i.e. an old device
that lacks support for key attestation.

NB (2021-10-30): THIS PATCH IS NO LONGER RECOMMENDED due to unavoidable
issues with this approach of blocking attestation from the keystore
service itself, and it is no longer sufficient by itself because
SafetyNet is now enforcing attestation based on device info. Please use
the frameworks/base patch instead [1].

[1] https://github.com/ProtonAOSP/android_frameworks_base/commit/13dc7d28c12d

Change-Id: I2dccd18f078eb7bd28c8e5ef265a97e9053c684e
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>